backend/requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
amqp==5.3.1
annotated-types==0.7.0
arrow==1.4.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.4
blessed==1.23.0
celery==5.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
croniter==6.0.0
Django==5.2.8
django-cors-headers==4.9.0
django-picklefield==3.3
django-q==1.3.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
et_xmlfile==2.0.0
h11==0.16.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonpath==0.82.2
jsonpath-ng==1.8.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
ply==3.11
prompt_toolkit==3.0.52
pyasn1==0.6.1
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
PySocks==1.7.1
pytest==9.0.1
pytest-result-log==1.2.2
pytest-xlsx==1.0.0
pytest-yaml==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==3.5.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
selenium==4.38.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tabulate==0.9.0
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websocket-client==1.9.0
wsproto==1.3.1
gunicorn==20.1.0
psycopg2-binary==2.9.9
